fix: support GPT-5 Responses API for OpenAI providers
